Edit Product Category has added to the project.
Includes:
On branch dev
Changes to be committed:
	modified:   src/Presentation/TopraqShop.Presentation.ServiceHost/Areas/Administration/Pages/Shared/_AdminLayout.cshtml
	deleted:    src/Presentation/TopraqShop.Presentation.ServiceHost/Areas/Administration/Pages/ShopManagement/ProductCategories/Create.cshtml.cs
	new file:   src/Presentation/TopraqShop.Presentation.ServiceHost/Areas/Administration/Pages/ShopManagement/ProductCategories/Edit.cshtml
	modified:   src/Presentation/TopraqShop.Presentation.ServiceHost/Areas/Administration/Pages/ShopManagement/ProductCategories/Index.cshtml
	modified:   src/Presentation/TopraqShop.Presentation.ServiceHost/Areas/Administration/Pages/ShopManagement/ProductCategories/Index.cshtml.cs
	new file:   src/Presentation/TopraqShop.Presentation.ServiceHost/wwwroot/AdminTheme/assets/datatables/dataTables.select.min.js
	modified:   src/Presentation/TopraqShop.Presentation.ServiceHost/wwwroot/AdminTheme/assets/js/site.js
	modified:   src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Concrete/CreateProductCategory.cs
	modified:   src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Concrete/EditProductCategory.cs
	modified:   src/ShopManagement/Domain/Application/TopraqShop.ShopManagement.Domain.Application.Contracts/ProductCategory/Concrete/ProductCategoryViewModel.cs
	modified:   src/ShopManagement/Domain/TopraqShop.ShopManagement.Domain.Base/ProductCategoryAggregate/IProductCategoryRepository.cs
	modified:   src/ShopManagement/Domain/TopraqShop.ShopManagement.Domain.Base/ProductCategoryAggregate/ProductCategoryBase.cs
	modified:   src/ShopManagement/Infrastructure/TopraqShop.ShopManagement.Infrastructure.EfCore/Mapping/ProductCategoryMapping.cs
	renamed:    src/ShopManagement/Infrastructure/TopraqShop.ShopManagement.Infrastructure.EfCore/Migrations/20211007160532_init.Designer.cs -> src/ShopManagement/Infrastructure/TopraqShop.ShopManagement.Infrastructure.EfCore/Migrations/20211009104847_init.Designer.cs
	renamed:    src/ShopManagement/Infrastructure/TopraqShop.ShopManagement.Infrastructure.EfCore/Migrations/20211007160532_init.cs -> src/ShopManagement/Infrastructure/TopraqShop.ShopManagement.Infrastructure.EfCore/Migrations/20211009104847_init.cs
	modified:   src/ShopManagement/Infrastructure/TopraqShop.ShopManagement.Infrastructure.EfCore/Migrations/ShopContextModelSnapshot.cs
	modified:   src/ShopManagement/Infrastructure/TopraqShop.ShopManagement.Infrastructure.EfCore/Repositories/ProductCategoryRepository.cs

+++ src/ShopManagement/Domain/TopraqShop.ShopManagement.Domain.Base/ProductCategoryAggregate/IProductCategoryRepository.cs
@@ -3,7 +3,7 @@ using TopraqShop.Framework.Base.Domain;
 
 namespace TopraqShop.ShopManagement.Domain.Base.ProductCategoryAggregate
 {
-    public interface IProductCategoryRepository:IRepository<byte, ProductCategoryBase>
+    public interface IProductCategoryRepository:IRepository<int, ProductCategoryBase>
     {
     }
 }
